+++ src/node/tests/test_schema.py
@@ -1,9 +1,15 @@
 from node import schema
 from node.base import BaseNode
 from node.behaviors import Schema
+from node.behaviors import SchemaAsAttributes
+from node.behaviors import SchemaAttributes
 from node.behaviors.schema import scope_field
+from node.interfaces import IAttributes
+from node.interfaces import INodeAttributes
 from node.interfaces import ISchema
+from node.interfaces import ISchemaAsAttributes
 from node.tests import NodeTestCase
+from node.utils import AttributeAccess
 from node.utils import UNSET
 from plumber import plumbing
 import uuid
@@ -33,6 +39,11 @@ class TestSchema(NodeTestCase):
         field.reset_scope()
         self.assertEqual(field.name, None)
         self.assertEqual(field.parent, None)
+
+    def test_Bool(self):
+        field = schema.Bool()
+        self.assertTrue(field.validate(True))
+        self.assertFalse(field.validate(1))
 
     def test_Int(self):
         field = schema.Int()
@@ -79,18 +90,6 @@ class TestSchema(NodeTestCase):
         self.assertTrue(field.validate(uuid.uuid4()))
         self.assertFalse(field.validate('1234'))
 
-    def test_Node(self):
-        field = schema.Node()
-        self.assertTrue(field.validate(BaseNode()))
-        self.assertFalse(field.validate(object()))
-
-        parent = BaseNode()
-        field.set_scope('name', parent)
-        node = field.deserialize('some-non-node-data')
-        self.assertIsInstance(node, BaseNode)
-        self.assertEqual(node.name, 'name')
-        self.assertTrue(node.parent is parent)
-
     def test_scope_field(self):
         field = schema.Field()
         parent = object()
@@ -114,20 +113,95 @@ class TestSchema(NodeTestCase):
         self.assertTrue(ISchema.providedBy(node))
 
         node['any'] = 'foo'
-        self.assertRaises(ValueError, node.__setitem__, 'int', '1')
+        with self.assertRaises(ValueError):
+            node['int'] = '1'
         node['int'] = 0
 
         self.checkOutput("""
         <class 'node.tests.test_schema.SchemaNode'>: None
-          any: 'foo'
-          int: 0
-        """, node.treerepr())
+        __any: 'foo'
+        __int: 0
+        """, node.treerepr(prefix='_'))
 
         self.assertEqual(node['any'], 'foo')
         self.assertEqual(node['int'], 0)
         self.assertEqual(node['float'], 1.)
-        self.assertRaises(KeyError, node.__getitem__, 'str')
+        with self.assertRaises(KeyError):
+            node['str']
 
         node.schema['*'] = schema.UUID()
-        self.assertRaises(ValueError, node.__setitem__, 'arbitrary', '1234')
+        with self.assertRaises(ValueError):
+            node['arbitrary'] = '1234'
         node['arbitrary'] = uuid.uuid4()
+
+    def test_SchemaAsAttributes(self):
+        @plumbing(SchemaAsAttributes)
+        class SchemaAsAttributesNode(BaseNode):
+            schema = {
+                'int': schema.Int(),
+                'float': schema.Float(default=1.),
+                'str': schema.Str(),
+                'bool': schema.Bool()
+            }
+
+        node = SchemaAsAttributesNode()
+        self.assertTrue(IAttributes.providedBy(node))
+        self.assertTrue(ISchemaAsAttributes.providedBy(node))
+
+        attrs = node.attrs
+        self.assertTrue(INodeAttributes.providedBy(attrs))
+        self.assertIsInstance(attrs, SchemaAttributes)
+        self.assertEqual(sorted(iter(attrs)), ['bool', 'float', 'int', 'str'])
+
+        attrs['int'] = 1
+        attrs['float'] = 2.
+        attrs['str'] = u'foo'
+        attrs['bool'] = True
+        self.assertEqual(node.storage, {
+            'float': 2.,
+            'int': 1,
+            'str': u'foo',
+            'bool': True
+        })
+        del attrs['bool']
+        with self.assertRaises(ValueError):
+            attrs['str'] = 1
+        with self.assertRaises(KeyError):
+            attrs['other']
+        with self.assertRaises(KeyError):
+            attrs['other'] = 'value'
+        with self.assertRaises(KeyError):
+            del attrs['other']
+
+        node.attribute_access_for_attrs = True
+        attrs = node.attrs
+        self.assertIsInstance(attrs, AttributeAccess)
+        self.assertEqual(attrs.int, 1)
+        self.assertEqual(attrs.float, 2.)
+        self.assertEqual(attrs.str, u'foo')
+
+        attrs.int = 2
+        attrs.float = 3.
+        attrs.str = u'bar'
+        self.assertEqual(node.storage, {'float': 3., 'int': 2, 'str': u'bar'})
+
+        child = node['child'] = BaseNode()
+        self.checkOutput("""
+        <class 'node.tests.test_schema.SchemaAsAttributesNode'>: None
+        __<class 'node.base.BaseNode'>: child
+        """, node.treerepr(prefix='_'))
+
+        self.assertEqual(list(iter(node)), ['child'])
+        self.assertTrue(node['child'] is child)
+        del node['child']
+
+        self.checkOutput("""
+        <class 'node.tests.test_schema.SchemaAsAttributesNode'>: None
+        """, node.treerepr(prefix='_'))
+
+        with self.assertRaises(KeyError):
+            node['int'] = 1
+        with self.assertRaises(KeyError):
+            node['int']
+        with self.assertRaises(KeyError):
+            del node['int']